Created a levels module, uses 'main.py' to test it
The levels module just fills out the background of the level. Random generation will be used to handle placement of enemies, traps, etc. Puzzles will be hardcoded in place for each level.

The levels sole function takes in images for the floor, floor edges, floor corners, and walls, and draws them on the screen to make a single level. Call this function within the game loop, boom, you've got yourself a level.