Basic Knockback and Invincibility
Enemies no longer die upon damaging the player, but instead knock the player back and give them invincibility frames.
To govern the above, the player class now has 4 new variables: a knockback timer, invincibility timer, knockback power in the X direction, and knockback power in the Y direction.
Knockback direction is determined by where the player was relative to the enemy. Currently, diagonal knockback is possible -- may want to be changed? You also can't move while being knocked back.
Player invincibility is not visually shown as of yet. Preferably, the player would flash in and out of visibility when invincible, but I don't know how to code that yet.